refactor: rename docling_extractor to pdf_extractor (#308)
The module uses PyMuPDF (fitz), not Docling. Rename to match the
actual implementation and update all imports and test patches.
